docs/requirements.txt
Jinja2<3.1
sphinx
pydata-sphinx-theme
matplotlib
sphinx-design
Pillow